moved all logging includes into cpp files
All logging includes are no in cpp files, thus the macro does not pollute
public include files. If the logging include is in public include files
the NO_ROS_LOGGING macro must also be defined in the application which
uses the library.

Further a NO_LOGGING macro is defined to completely disable logging